sistemas pedidos/backend/requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
psycopg2-binary>=2.9.3
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
requests>=2.20.0
